+++ src/main/java/gregicality/multiblocks/api/metatileentity/GCYMRecipeMapMultiblockController.java
@@ -15,8 +15,10 @@ import org.jetbrains.annotations.Nullable;
 import gregtech.api.GTValues;
 import gregtech.api.metatileentity.ITieredMetaTileEntity;
 import gregtech.api.metatileentity.multiblock.MultiMapMultiblockController;
+import gregtech.api.metatileentity.multiblock.MultiblockDisplayText;
 import gregtech.api.pattern.TraceabilityPredicate;
 import gregtech.api.recipes.RecipeMap;
+import gregtech.api.util.GTUtility;
 
 import gregicality.multiblocks.api.capability.IParallelMultiblock;
 import gregicality.multiblocks.api.capability.impl.GCYMMultiblockRecipeLogic;
@@ -45,14 +47,26 @@ public abstract class GCYMRecipeMapMultiblockController extends MultiMapMultiblo
 
     @Override
     protected void addDisplayText(List<ITextComponent> textList) {
-        super.addDisplayText(textList);
-        List<ITieredMetaTileEntity> list = getAbilities(GCYMMultiblockAbility.TIERED_HATCH);
-        if (GCYMConfigHolder.globalMultiblocks.enableTieredCasings && !list.isEmpty()) {
-            long maxVoltage = Math.min(GTValues.V[list.get(0).getTier()],
-                    Math.max(energyContainer.getInputVoltage(), energyContainer.getOutputVoltage()));
-            String voltageName = GTValues.VNF[list.get(0).getTier()];
-            textList.add(new TextComponentTranslation("gcym.multiblock.tiered_hatch.tooltip", maxVoltage, voltageName));
-        }
+        MultiblockDisplayText.builder(textList, isStructureFormed())
+                .setWorkingStatus(recipeMapWorkable.isWorkingEnabled(), recipeMapWorkable.isActive())
+                .addEnergyUsageLine(getEnergyContainer())
+                .addEnergyTierLine(GTUtility.getTierByVoltage(recipeMapWorkable.getMaxVoltage()))
+                .addCustom(tl -> {
+                    // Tiered Hatch Line
+                    if (isStructureFormed()) {
+                        List<ITieredMetaTileEntity> list = getAbilities(GCYMMultiblockAbility.TIERED_HATCH);
+                        if (GCYMConfigHolder.globalMultiblocks.enableTieredCasings && !list.isEmpty()) {
+                            long maxVoltage = Math.min(GTValues.V[list.get(0).getTier()],
+                                    Math.max(energyContainer.getInputVoltage(), energyContainer.getOutputVoltage()));
+                            String voltageName = GTValues.VNF[list.get(0).getTier()];
+                            tl.add(new TextComponentTranslation("gcym.multiblock.tiered_hatch.tooltip", maxVoltage,
+                                    voltageName));
+                        }
+                    }
+                })
+                .addParallelsLine(recipeMapWorkable.getParallelLimit())
+                .addWorkingStatusLine()
+                .addProgressLine(recipeMapWorkable.getProgressPercent());
     }
 
     @Override