+++ src/main/java/gregicality/multiblocks/api/unification/material/GCYMMaterialIconTypes.java
@@ -1,0 +1,10 @@
+package gregicality.multiblocks.api.unification.material;
+
+import gregtech.api.unification.material.info.MaterialIconType;
+
+public final class GCYMMaterialIconTypes {
+
+    public static final MaterialIconType molten = new MaterialIconType("molten");
+
+    private GCYMMaterialIconTypes() {}
+}