+++ src/main/java/gregicality/multiblocks/api/unification/properties/AlloyBlastPropertyAddition.java
@@ -2,6 +2,7 @@ package gregicality.multiblocks.api.unification.properties;
 
 import java.util.List;
 
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
 import gregtech.api.GregTechAPI;
@@ -12,9 +13,10 @@ import gregtech.api.unification.stack.MaterialStack;
 
 import gregicality.multiblocks.api.unification.GCYMMaterialFlags;
 
-public final class GCYMLatePropertyAddition {
+@ApiStatus.Internal
+public final class AlloyBlastPropertyAddition {
 
-    private GCYMLatePropertyAddition() {}
+    private AlloyBlastPropertyAddition() {}
 
     public static void init() {
         for (Material material : GregTechAPI.materialManager.getRegisteredMaterials()) {
@@ -24,7 +26,7 @@ public final class GCYMLatePropertyAddition {
         }
     }
 
-    public static void addAlloyBlastProperty(@NotNull Material material) {
+    private static void addAlloyBlastProperty(@NotNull Material material) {
         final List<MaterialStack> components = material.getMaterialComponents();
         // ignore materials which are not alloys
         if (components.size() < 2) return;
@@ -35,7 +37,7 @@ public final class GCYMLatePropertyAddition {
         if (!material.hasProperty(PropertyKey.FLUID)) return;
 
         // if there are more than 2 fluid-only components in the material, do not generate a hot fluid
-        if (components.stream().filter(GCYMLatePropertyAddition::isMaterialStackFluidOnly).limit(3).count() > 2) {
+        if (components.stream().filter(AlloyBlastPropertyAddition::isMaterialStackFluidOnly).limit(3).count() > 2) {
             return;
         }