+++ src/main/java/gregicality/multiblocks/common/GCYMEventHandlers.java
@@ -4,14 +4,18 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
+import org.jetbrains.annotations.ApiStatus;
+
 import gregtech.api.unification.material.event.MaterialEvent;
 import gregtech.api.unification.material.event.PostMaterialEvent;
 
 import gregicality.multiblocks.GregicalityMultiblocks;
+import gregicality.multiblocks.api.fluids.GeneratedFluidHandler;
 import gregicality.multiblocks.api.unification.GCYMMaterialFlagAddition;
 import gregicality.multiblocks.api.unification.GCYMMaterials;
-import gregicality.multiblocks.api.unification.properties.GCYMLatePropertyAddition;
+import gregicality.multiblocks.api.unification.properties.AlloyBlastPropertyAddition;
 
+@ApiStatus.Internal
 @Mod.EventBusSubscriber(modid = GregicalityMultiblocks.MODID)
 public final class GCYMEventHandlers {
 
@@ -24,7 +28,8 @@ public final class GCYMEventHandlers {
 
     @SubscribeEvent
     public static void registerMaterialsPost(PostMaterialEvent event) {
-        GCYMLatePropertyAddition.init();
+        AlloyBlastPropertyAddition.init();
         GCYMMaterialFlagAddition.initLate();
+        GeneratedFluidHandler.init();
     }
 }